fix(backend): inactivity auto-pause and Carbon 3 signed-diff inversions [skip ci]
games:monitor-inactivity (00:48):
- pauseGame() accepts a null pauser (system pause); the monitor pauses
  with the server's last known clocks instead of passing 'inactivity'
  as the user id (TypeError, so auto-pause never fired)
- exclude computer/synthetic bot games from both monitor branches;
  games:cleanup-abandoned aborts them at 1 h with '*'
- Game::getInactiveSeconds()/getPausedSeconds() return absolute diffs

Carbon 3 diff audit (01:18):
- GameRoomService::requestResume stale cleanup measured
  resume_requested_at -> now(); the 20-min/1-min branches never fired
  and the same-user 10 s branch matched at any age
- ChampionshipGameTimeoutService needs_warning no longer true for every
  upcoming match
- CheckExpiredMatchesJob forfeit comparison no longer flipped (path is
  still unreachable via $game->moves(); fixed separately)
- (int)-cast float diffs reaching int params, JSON countdowns and
  log-rotation day counts

Tests: MonitorGameInactivityTest (7), ResumeRequestStaleCleanupTest (4),
needs_warning unit test. Full suite on the committed tree: 368 passed /
158 skipped / 1 risky / 0 failures.

+++ chess-backend/app/Http/Controllers/WebSocketController.php
@@ -1087,7 +1087,7 @@ class WebSocketController extends Controller
             }
 
             if ($existing) {
-                $secondsRemaining = max(0, now()->diffInSeconds($existing->expires_at, false));
+                $secondsRemaining = max(0, (int) now()->diffInSeconds($existing->expires_at, false));
 
                 return response()->json([
                     'message' => 'Resume request already pending.',
@@ -1113,7 +1113,7 @@ class WebSocketController extends Controller
                     'message' => 'Resume request cooldown active.',
                     'cooldown' => true,
                     'can_request_again_at' => $retryAt->toIso8601String(),
-                    'can_request_again_in_seconds' => max(0, now()->diffInSeconds($retryAt, false)),
+                    'can_request_again_in_seconds' => max(0, (int) now()->diffInSeconds($retryAt, false)),
                 ], 429); // Too Many Requests
             }
